Update: login bloc, update: login page.
Update login bloc with RxDart, update the streamController to behaviorSubject that because the combine latest doens't work with stream controller, create a new validation formValidStream to listen if the email and password are correct.
Update login page with a new stream builder into the elevated button to listen if the email and password are correct, unlock the button.