Update: login bloc, update: login page update; home page.
Update login bloc page with the get of the email and password controller value.
update login page with a new method to print the email and password in the console and a navigator to the home page if the email and password has data.
update home page with the title of the app bar and a column that shows the email and password entered by the user.