Update: producto page.
Update producto page with a new method to show a Snack bar message when the item is saved.
Create a new variable to disable the button when the item is save.
Use the navigator pop when the user tap the save button the will back to the home page automatically.