Update: Home page, update: producto page, update: productos provider.
Update the home page with the new widget Card with a Column as a child, as a children of the column it will show the image if the user picked one if not it will show an asset image, each image has a ListTile with the name of the producto, its price and the Switch list tile making the producto available or not.
Update the producto page with a new code that shows the product image into the producto page.