Add: preferencias usuario, update: usuario provider, update: main.
Create the class preferencias usuario to save the user preferences with the get and set of the token.
Update usuario provider with the instance of preferencias usuario to save the idToken generated in firebase into the device.
update main with the instance of preferencias usuario, including the async to get the prefs.