Add: Productos bloc, update: provider, add: Registro page, add: preferencias usuario, add: usuario provider, update: utils.
Create the class productos bloc to manage the streams.
Update provider creating an instance of Productos bloc to create a new Inherited Widget to search into the context the instances of productos bloc.
Create registro page which is similar to login page with a few changes into it.
Create preferencias usuario to save the token in the internal storage of the device.
Create usuario provider to connect with the database to save the new users and the idToken and validate if the token is still available.
Update utils to create an error message to show it to the user if the password is incorrect or if the email exists.