Update: producto page, update: productos provider, update: main
Update producto page with a conditional in the method mostrarFoto to show producto.fotoUrl but if the user selects another image from the gallery or from camera the method will show that image.
Update productos provider with a queryParameters that contains the token to get the access to the app.
Update main changing the title of the app.